+++ tests/test_server.rs
@@ -449,9 +449,10 @@ async fn handle_or_drop_publish_after_disconnect(
     .unwrap();
     io.encode(codec::Packet::Disconnect, &codec).unwrap();
     io.flush(true).await.unwrap();
+    sleep(Millis(1750)).await;
+    io.close();
     drop(io);
-
-    sleep(Millis(50)).await;
+    sleep(Millis(500)).await;
 
     assert!(disconnect.load(Relaxed));
 
@@ -770,7 +771,7 @@ async fn test_frame_read_rate() -> std::io::Result<()> {
     sleep(Millis(1000)).await;
     assert!(!check.load(Relaxed));
 
-    sleep(Millis(2100)).await;
+    sleep(Millis(2300)).await;
     assert!(check.load(Relaxed));
 
     Ok(())