+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [2.0.0] - 2024-04-1x
 
+* Mark `Control` type as `non exhaustive`
+
 * Rename `ControlMessage` to `Control`
 
 * Remove protocol variant services